[Backend] Partial API Implementation 📄

+++ db/seller.sql.go
@@ -61,7 +61,7 @@ LIMIT $2
 
 type SellerBestSellProductParams struct {
 	SellerName string             `json:"seller_name" param:"seller_name"`
-	Limit      int32              `json:"limit"`
+	Limit      int64              `json:"limit"`
 	Time       pgtype.Timestamptz `json:"time" swaggertype:"string"`
 }
 
@@ -248,8 +248,8 @@ LIMIT $2 OFFSET $3
 
 type SellerGetCouponParams struct {
 	SellerName string `json:"seller_name" param:"seller_name"`
-	Limit      int32  `json:"limit"`
-	Offset     int32  `json:"offset"`
+	Limit      int64  `json:"limit"`
+	Offset     int64  `json:"offset"`
 }
 
 type SellerGetCouponRow struct {
@@ -424,8 +424,8 @@ LIMIT $2 OFFSET $3
 
 type SellerGetOrderParams struct {
 	SellerName string `json:"seller_name" param:"seller_name"`
-	Limit      int32  `json:"limit"`
-	Offset     int32  `json:"offset"`
+	Limit      int64  `json:"limit"`
+	Offset     int64  `json:"offset"`
 }
 
 type SellerGetOrderRow struct {
@@ -980,8 +980,8 @@ LIMIT $2 OFFSET $3
 
 type SellerProductListParams struct {
 	SellerName string `json:"seller_name" param:"seller_name"`
-	Limit      int32  `json:"limit"`
-	Offset     int32  `json:"offset"`
+	Limit      int64  `json:"limit"`
+	Offset     int64  `json:"offset"`
 }
 
 type SellerProductListRow struct {
@@ -1063,7 +1063,7 @@ LIMIT $3
 type SellerSearchTagParams struct {
 	SellerName string `json:"seller_name" param:"seller_name"`
 	Name       string `json:"name"`
-	Limit      int32  `json:"limit"`
+	Limit      int64  `json:"limit"`
 }
 
 type SellerSearchTagRow struct {
